Speed up parsing significantly
Passing the Fields interface into the readX functions results
in the passed-in struct being copied, creating an object that
needs to be cleaned up by GC later.

This results in a large part of program runtime being spent on
GC, as this function is used in many hot functions, such as
createStopTime().

According to pprof, removing these copies speeds up the entire
program by about 14%, and reduces the amount of garbage objects generated
by over 25%.