Don't return entire StopTime object from createStopTime
The sole caller only uses one field, and by returning this field
directly we allow Go's escape analysis to do it's thing and avoid
a costly heap allocation. According to pprof, this cuts out
several million allocations when parsing the German GTFS dataset.

This function is not exported, therefore this change does not alter the public API.